requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
Flask-SocketIO==5.3.6
google-generativeai==0.8.3
python-dotenv==1.1.1
werkzeug==3.1.3
gunicorn==23.0.0
requests==2.31.0
notion-client==2.2.1
